OrderArchiveHandler and UI remade

+++ src/model/Pizza.java
@@ -3,9 +3,9 @@ package model;
 public class Pizza {
     private String name;
     private String ingredients;
-    private int price;
+    private double price;
 
-    public Pizza(String name, String ingredients, int price){
+    public Pizza(String name, String ingredients, double price){
         this.name = name;
         this.ingredients = ingredients;
         this.price = price;
@@ -13,6 +13,6 @@ public class Pizza {
 
     public String getName() {return name;}
     public String getIngredients() {return ingredients;}
-    public int getPrice() {return price;}
+    public double getPrice() {return price;}
 
 }
